backend/requirements.txt
# =============================================================================
# Nature Marketplace - Backend Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Django
# -----------------------------------------------------------------------------
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9,<3.0  # PostgreSQL adapter
dj-database-url>=2.1,<3.0  # Database URL parsing

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
djangorestframework-simplejwt>=5.3,<6.0  # JWT authentication
django-allauth>=0.61,<1.0  # OAuth & social authentication
dj-rest-auth>=5.0,<6.0  # REST API for allauth
django-ratelimit>=4.1,<5.0  # Rate limiting
bleach>=6.1,<7.0  # HTML sanitization
cryptography>=41.0,<43.0  # Required for allauth social providers
PyJWT>=2.8,<3.0  # JWT encoding/decoding

# -----------------------------------------------------------------------------
# Object Storage (MinIO/S3)
# -----------------------------------------------------------------------------
django-storages>=1.14,<2.0  # Storage backends
boto3>=1.34,<2.0  # S3-compatible client

# -----------------------------------------------------------------------------
# Stripe Integration
# -----------------------------------------------------------------------------
stripe>=7.0,<8.0  # Stripe Python SDK

# -----------------------------------------------------------------------------
# API Documentation
# -----------------------------------------------------------------------------
drf-spectacular>=0.27,<1.0  # OpenAPI schema generation

# -----------------------------------------------------------------------------
# Environment & Configuration
# -----------------------------------------------------------------------------
python-dotenv>=1.0,<2.0  # Environment variable loading
python-decouple>=3.8,<4.0  # Settings management

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
Pillow>=10.0,<11.0  # Image processing
python-dateutil>=2.8,<3.0  # Date utilities
requests>=2.31,<3.0  # HTTP client

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0  # Test fixtures

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
flake8>=6.1,<7.0
black>=23.12,<24.0
isort>=5.13,<6.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0  # Django type stubs
djangorestframework-stubs>=3.14,<4.0  # DRF type stubs

# -----------------------------------------------------------------------------
# Production
# -----------------------------------------------------------------------------
gunicorn>=21.2,<22.0  # WSGI server
whitenoise>=6.6,<7.0  # Static file serving
